simplified styling and folder
