+++ 01-Login/EXAMPLE.env
@@ -1,0 +1,2 @@
+PASSAGE_APP_ID=O8xAAilHFn16nOrfsR1LwWAT
+PASSAGE_API_KEY=2hpPMYOpje.LviN6hARr2s8TgTlH8cUS9P4gc3h7Gm1WGIVXCLFgOWKtw2isCni22cefCnythVc
