+++ 01-Login/pages/index.js
@@ -1,0 +1,21 @@
+import { useEffect } from 'react';
+
+export default function Home({appID}) {
+  useEffect(()=>{
+    require('@passageidentity/passage-auth');
+  }, []);
+
+  return (
+    <>
+    <passage-auth app-id={appID}></passage-auth>
+    </>
+  )
+}
+
+export async function getStaticProps(){
+  return {
+    props: {
+      appID: process.env.PASSAGE_APP_ID
+    }
+  };
+}
