+++ 01-Login/styles/Banner.module.css
@@ -1,0 +1,28 @@
+.mainHeader{
+    padding: 20px 30px;
+    display: flex;
+    align-items: center;
+    background-color: #27417E;
+    color: white;
+}
+.header-text {
+    font-size: 24px;
+    margin-left: 10px;
+}
+    
+.passageLogo {
+    background-image: url('https://storage.googleapis.com/customer-assets-prod-ycfxsoekhrvsmkgvonyymmic/blog-passage-svg/passage-logo.svg');
+    background-repeat: no-repeat;
+    width: 60px;
+    height: 60px;
+    cursor: pointer;
+}
+.spacer {
+    flex-grow: 1;
+}
+    
+.link {
+    margin-left: 20px;
+    color: white;
+    text-decoration-color: white;
+}
