--- EXAMPLE.env
@@ -1,2 +1,0 @@
-PASSAGE_APP_ID=
-PASSAGE_API_KEY=
