--- styles/Authorized.module.css
@@ -1,116 +1,0 @@
-.bodyClass{
-    background-repeat: no-repeat;
-    background-image: 
-      url('/fingerprint-left-success.svg'), 
-      url('/fingerprint-right-success.svg');
-    background-position: 
-      top 130px left -50%,
-      top 300px right -50%;
-    background-size: 50%;
-  }
-
-.mainContainer{
-    padding-top: 60px;
-    padding-left: 75px;
-    padding-right: 75px;
-    padding-bottom: 60px;
-    display: flex;
-    flex-direction: column;
-    color: #263768
-}
-
-.spacer{
-    flex-grow: 1;
-}
-
-.title {
-    font-size: 64px;
-    font-weight: 700;
-}
-
-.message {
-    font-size: 36px;
-}
-
-.subMessage {
-    font-size: 24px;
-    align-self: center;
-    padding-bottom: 10px;
-}
-
-.imgContainer {
-    height: 310px;
-    align-self: center;
-}
-
-.imgContainer > img {
-    max-width: 100%;
-    max-height: 100%;
-}
-
-/* for shorter 13" dekstop screens */
-@media screen and (max-height: 845px) {
-
-    .title {
-        font-size: 48px;
-    }
-    
-    .message {
-        font-size: 24px;
-    }
-    
-    .subMessage {
-        font-size: 20px;
-    }
-
-    .main-footer{
-        min-height: 45px;
-    }
-}
-
-/* mobile rules begin here */
-@media screen and (max-width: 760px){
-    .bodyClass{
-        background-position: 
-          top 100px left -120px,
-          top 340px right -120px;
-        background-size: 70%;
-      }
-
-    .mainContainer{
-        padding-top: 45px;
-        padding-left: 30px;
-        padding-right: 30px;
-    }
-
-    .title {
-        font-size: 36px;
-    }
-    
-    .message {
-        font-size: 18px;
-    }
-    
-    .subMessage {
-        font-size: 16px;
-        align-self: center;
-    }
-}
-
-@media screen and (max-width: 450px){
-    .bodyClass{
-        background-position: 
-          top 100px left -140px,
-          top 450px right -120px;
-        background-size: 100%;
-      }    
-    .imgContainer {
-        height: 260px;
-    }
-}
-
-@media screen and (max-height: 765px){
-    .bodyClass{
-        height: calc(100vh + 45px);
-    }
-}
