--- styles/Banner.module.css
@@ -1,63 +1,0 @@
-.mainHeader{
-    display: flex;
-    align-items: flex-start;
-    padding-top: 20px;
-    color: #263768;
-    position: relative;
-  }
-  
-.passageLogo {
-    background-image: url('/passage-full-logo.svg');
-    background-repeat: no-repeat;
-    width: 300px;
-    height: 60px;
-    margin-left: 40px;
-    cursor: pointer;
-}
-
-.spacer {
-    flex-grow: 1;
-} 
-  
-.home {
-    display: flex;
-    align-items: center;
-    margin-right: 20px;
-  }
-  
- .links {
-    display: flex;
-    color: #263768;
-    text-decoration-color: #263768;
-    text-decoration: underline;
-    cursor: pointer;
-    align-items: center;
-  }
-  
- .text {
-    margin-right: 10px;
-  }
-  
-.icon {
-    background-image: url('/home.svg');
-    background-repeat: no-repeat;
-    width: 28px;
-    height: 26px;
-  }
-
-.logout {
-    text-decoration: underline;
-    user-select: none;
-    cursor: pointer;
-    margin-right: 20px;
-}
-
-@media screen and (max-width: 530px) {
-    .passageLogo {
-      background-image: url('/passage-logo.svg');
-      background-repeat: no-repeat;
-      width: 50px;
-      height: 60px;
-      margin-left: 30px;
-    }
-  }
