--- styles/Home.module.css
@@ -1,131 +1,0 @@
-.bodyClass{
-  background-repeat: no-repeat;
-  background-image: 
-    url('/fingerprint-left-index.svg'), 
-    url('/fingerprint-right-index.svg');
-  background-position: 
-    top 160px left -50%,
-    top 300px right -50%;
-  background-size: 50%;
-}
-
-.mainContainer {
-  display: flex;
-}
-
-.spacer{
-  flex-grow: 1;
-}
-
-.earlyAccessContainer{
-  min-width: 325px;
-  width: 35%;
-  color: white;
-  background-color: #27417E;
-  border-top-left-radius: 30px;
-  border-bottom-left-radius: 30px;
-  display: flex;
-  flex-direction: column;
-}
-
-.title {
-  font-weight: 700;
-  font-size: 48px;
-  margin-top: 65px;
-  padding-left: 45px;
-  padding-right: 65px;
-}
-
-.bodyText{
-  padding-left: 45px;
-  padding-right: 45px;
-  padding-bottom: 35px;
-  font-size: 24px;
-}
-
-.authContainer{
-  width: 65%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  padding-left: 20px;
-  padding-right: 20px;
-}
-
-/* for shorter 13" dekstop screens */
-@media screen and (max-height: 845px) {
-  .title {
-    font-size: 40px;
-  }
-  .bodyText{
-    font-size: 20px;
-  }
-}
-
-/* steps to mobile rules begin here */
-@media screen and (max-width: 870px) {
-  .bodyClass{
-    background-position: 
-      top 110px left -120px,
-      top 300px right -120px;
-    background-size: 70%;
-  }
-
-  .mainContainer{
-    flex-direction: column;
-    width: 100%;
-  }
-
-  .earlyAccessContainer{
-    min-width: unset;
-    width: 100%;
-    min-height: 30%;
-    border-top-left-radius: 30px;
-    border-top-right-radius: 30px;
-    border-bottom-left-radius: 0px;
-    border-bottom-right-radius: 0px;
-  }
-
-  .title {
-    font-size: 32px;
-    margin-top: 30px;
-    padding-left: 30px;
-    padding-right: 30px;
-  }
-  .bodyText{
-    font-size: 20px;
-    padding-left: 30px;
-    padding-right: 30px;
-    padding-bottom: 15px;
-  }
-
-  .authContainer{
-    width: 100%;
-    height: 100%;
-    flex-grow: 1;
-    padding-top: 20px;
-    padding-left: 0px;
-    padding-right: 0px;
-  }
-}
-
-@media screen and (max-width: 670px) {
-  .bodyClass{
-    background-position: 
-      top 110px left -120px,
-      top 450px right -120px;
-    background-size: 100%;
-  }
-  .title {
-    font-size: 24px;
-  }
-  .bodyText{
-    font-size: 16px;
-  }
-}
-
-@media screen and (max-width: 460px) {
-  .earlyAccessContainer{
-    min-height: unset;
-  }
-}
