--- styles/Unauthorized.module.css
@@ -1,113 +1,0 @@
-.bodyClass{
-    background-repeat: no-repeat;
-    background-image: 
-      url('/fingerprint-left-error.svg'), 
-      url('/fingerprint-right-error.svg');
-    background-position: 
-      top 300px left -50%,
-      top 120px right -50%;
-    background-size: 50%;
-}
-
-.mainContainer {
-    padding-top: 75px;
-    padding-left: 75px;
-    padding-right: 75px;
-    padding-bottom: 40px;
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-}
-
-.spacer{
-  flex-grow: 1;
-}
-
-.flexContainer{
-    display: flex;
-}
-
-.messageContainer {
-    width: 50%;
-    color: #263768;
-}
-
-.title {
-    font-size: 64px;
-    font-weight: 700;
-}
-
-.message {
-    margin-top: 10px;
-    font-size: 36px;
-}
-
-.returnHome{
-    text-align: center;
-    font-size: 24px;
-    font-weight: 400;
-    color: #5F5D5D;
-}
-
-.returnHome a{
-    color: #263768;
-    font-weight: 700;
-}
-
-.imgContainer {
-    height: 460px;
-}
-
-.imgContainer > img {
-    max-width: 100%;
-    max-height: 100%;
-}
-
-@media screen and (max-width: 660px){
-    .mainContainer{
-        padding-top: 45px;
-        padding-left: 30px;
-        padding-right: 30px;
-    }
-    .spacer{
-        flex-grow: 0;
-    }
-    .flexContainer{
-        flex-direction: column;
-        flex-grow: 1;
-    }
-    .messageContainer {
-        width: 100%;
-    }
-    .bodyImage{
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        align-self: center;
-        flex-grow: 1;
-    }
-    .imgContainer{
-        height: 300px;
-    }
-}
-
-@media screen and (max-width: 500px){
-    .bodyClass{
-        background-position: 
-          top 450px left -120px,
-          top 90px right -190px;
-        background-size: 100%;
-      }
-    .title {
-        font-size: 36px;
-        font-weight: 700;
-    }
-    
-    .message {
-        margin-top: 10px;
-        font-size: 18px;
-    }
-    .returnHome{
-        font-size: 16px;
-    }
-}
